improvement: use pkg/feature and remove remote.go
