+++ pkg/daemon/controller/remotesubnet.go
@@ -1,8 +1,9 @@
 package controller
 
 import (
+	"reflect"
+
 	networkingv1 "github.com/oecp/rama/pkg/apis/networking/v1"
-	"github.com/oecp/rama/pkg/utils"
 )
 
 func (c *Controller) enqueueAddOrDeleteRemoteSubnet(obj interface{}) {
@@ -14,7 +15,7 @@ func (c *Controller) enqueueUpdateRemoteSubnet(oldObj, newObj interface{}) {
 	newRs := newObj.(*networkingv1.RemoteSubnet)
 
 	if oldRs.Spec.ClusterName != newRs.Spec.ClusterName ||
-		!utils.Intersect(&oldRs.Spec.Range, &newRs.Spec.Range) ||
+		!reflect.DeepEqual(oldRs.Spec.Range, newRs.Spec.Range) ||
 		networkingv1.GetRemoteSubnetType(oldRs) != networkingv1.GetRemoteSubnetType(newRs) {
 		c.subnetQueue.Add(ActionReconcileSubnet)
 	}
