--- pkg/daemon/iptables/remote.go
@@ -1,18 +1,0 @@
-package iptables
-
-import (
-	"net"
-)
-
-func (mgr *Manager) RecordRemoteNodeIP(nodeIP net.IP) {
-	mgr.remoteNodeIPList = append(mgr.remoteNodeIPList, nodeIP)
-}
-
-func (mgr *Manager) RecordRemoteSubnet(subnetCidr *net.IPNet, isOverlay bool) error {
-	if isOverlay {
-		mgr.remoteOverlaySubnet = append(mgr.remoteOverlaySubnet, subnetCidr)
-	} else {
-		mgr.remoteUnderlaySubnet = append(mgr.remoteUnderlaySubnet, subnetCidr)
-	}
-	return nil
-}
